ai-ml/requirements.txt
# REGIQ AI/ML Dependencies
# Core AI/ML libraries and frameworks

# =============================================================================
# CORE ML FRAMEWORKS
# =============================================================================
torch>=2.0.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# =============================================================================
# NLP & LLM LIBRARIES
# =============================================================================
transformers>=4.30.0
tokenizers>=0.13.0
datasets>=2.12.0
spacy>=3.6.0
langchain>=0.0.200
llama-index>=0.7.0
sentence-transformers>=2.2.0
openai>=0.27.0
google-generativeai>=0.1.0
anthropic>=0.3.0

# =============================================================================
# FAIRNESS & EXPLAINABILITY
# =============================================================================
fairlearn>=0.9.0
aif360>=0.5.0
shap>=0.42.0
lime>=0.2.0
interpret>=0.4.0
alibi>=0.9.0

# =============================================================================
# PROBABILISTIC & SIMULATION
# =============================================================================
pymc>=5.0.0
arviz>=0.15.0
prophet>=1.1.0
statsmodels>=0.14.0
monte-carlo>=0.1.0

# =============================================================================
# DATA PROCESSING & STORAGE - Local First
# =============================================================================
sqlalchemy>=2.0.0
# SQLite (built into Python)
# PostgreSQL for production
psycopg2-binary>=2.9.0
# Optional: Redis for caching
redis>=4.5.0
# Local vector databases
chromadb>=0.4.0
faiss-cpu>=1.7.0
# Optional: Elasticsearch for advanced search
# elasticsearch>=8.8.0

# =============================================================================
# WEB FRAMEWORKS & APIs
# =============================================================================
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.31.0
httpx>=0.24.0

# =============================================================================
# VISUALIZATION & REPORTING
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0
reportlab>=4.0.0
weasyprint>=59.0

# =============================================================================
# LOCAL STORAGE & DEPLOYMENT
# =============================================================================
# Local file system storage (no cloud dependencies)
pathlib>=1.0.0
# Optional: Google Cloud Storage for future scaling
# google-cloud-storage>=2.9.0
docker>=6.1.0

# =============================================================================
# MONITORING & LOGGING
# =============================================================================
mlflow>=2.4.0
wandb>=0.15.0
prometheus-client>=0.17.0
sentry-sdk>=1.28.0

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
joblib>=1.3.0
dask>=2023.6.0

# =============================================================================
# TESTING & DEVELOPMENT
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
psutil>=5.9.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# =============================================================================
# JUPYTER & NOTEBOOKS
# =============================================================================
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=6.5.0

# =============================================================================
# SECURITY & COMPLIANCE
# =============================================================================
cryptography>=41.0.0
pyjwt>=2.7.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# =============================================================================
# PERFORMANCE & OPTIMIZATION
# =============================================================================
numba>=0.57.0
cython>=0.29.0
psutil>=5.9.0
memory-profiler>=0.60.0

# =============================================================================
# DOCUMENT PROCESSING
# =============================================================================
pypdf2>=3.0.0
python-docx>=0.8.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# =============================================================================
# TIME SERIES & FORECASTING
# =============================================================================
fbprophet>=0.7.1
pmdarima>=2.0.0
sktime>=0.21.0

# =============================================================================
# GRAPH & NETWORK ANALYSIS
# =============================================================================
networkx>=3.1.0
igraph>=0.10.0

# =============================================================================
# AUDIO & VISION (if needed)
# =============================================================================
pillow>=10.0.0
opencv-python>=4.8.0
librosa>=0.10.0

# =============================================================================
# ASYNC & CONCURRENCY
# =============================================================================
asyncio>=3.4.3
aiohttp>=3.8.0
celery>=5.3.0

# =============================================================================
# VERSION PINNING FOR STABILITY
# =============================================================================
# Pin critical versions for production stability
# Uncomment and adjust versions as needed for your environment

# torch==2.0.1
# tensorflow==2.13.0
# transformers==4.30.2
# scikit-learn==1.3.0
# pandas==2.0.3
# numpy==1.24.3
